backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
pymilvus==2.3.6
sentence-transformers==2.3.1
torch==2.7.0
transformers==4.37.2
httpx==0.26.0
aiohttp==3.9.1
python-dotenv==1.0.0
jieba==0.42.1
numpy==1.26.3
redis==5.0.1
ragas==0.1.12
datasets==2.18.0
langchain-community>=0.0.38